requirements.txt
yfinance>=0.2.58
pandas
python-dotenv
gspread
google-auth